modelling/requirements.txt
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
torch==2.4.1
torchvision==0.19.1
